Refactor and renamespace
Renamespaced classes for offence migrator and seeder
for clarity, as well as adding string addendums to
in-app StringExtension class.
