+++ db/offence_code_seeder.rb
@@ -1,0 +1,28 @@
+  class OffenceCodeSeeder
+    attr_reader :description, :class_letter
+
+    def initialize(description, class_letter)
+      @description = description
+      @class_letter = class_letter
+    end
+
+    def unique_code
+      modifier = 0
+      unique_code = code
+      unique_code = code(modifier += 1) while exists?(unique_code)
+      unique_code
+    end
+
+    private
+
+    def code(modifier = nil)
+      code = description.abbreviate +
+        modifier.to_s +
+        '_' +
+        class_letter
+    end
+
+    def exists?(code)
+      Offence.find_by(unique_code: code).present?
+    end
+  end
