+++ lib/data_migrator/offence_code_generator.rb
@@ -1,0 +1,14 @@
+class OffenceCodeGenerator
+  attr_reader :offence
+
+  def initialize(offence)
+    @offence = offence
+  end
+
+  def code(modifier = nil)
+    offence.description.abbreviate +
+      modifier.to_s +
+      '_' +
+      offence.offence_class.class_letter
+  end
+end
